move the Rule, Check and Expression builders to their own files
